add tests for the algorithm and service

+++ src/main/java/ru/butenko/model/User.java
@@ -1,24 +1,26 @@
 package ru.butenko.model;
 
+import jakarta.persistence.*;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
-import javax.persistence.*;
 
 @Entity
 @NoArgsConstructor
 @AllArgsConstructor
 @Builder
 @Data
-@Table(name="users")
+@Table(name = "users")
 public class User {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private long id;
-    @Column(name="user_name")
+    private Integer id;
+
+    @Column(name = "user_name")
     private String username;
+
     @Column
     private String user_password;
 }
